moving entitis dir to src level
